Made get_column(), get_row(), get_all_columns() and get_all_rows() methods.
get_column() and get_row() returns the column and row with the element objects.
get_column_values() and get_row_values() return the column and rows only with the values.
get_all_columns and get_all_rows return all the rows in the form of a list, which hold the element objects.